+++ compiler-tests/src/test/data/box/inject/member/MultiInheritanceMemberInject.kt
@@ -1,0 +1,25 @@
+// https://github.com/ZacSweers/metro/issues/649
+
+@Inject class Foo
+
+@Inject class Bar
+
+open class Parent {
+
+  @Inject lateinit var foo: Foo
+}
+
+class Child : Parent() {
+
+  @Inject lateinit var bar: Bar
+}
+
+@DependencyGraph(AppScope::class)
+interface TestGraph {
+  fun inject(target: Child)
+}
+
+fun box(): String {
+  // TODO()
+  return "OK"
+}