+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -487,6 +487,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
       }
 
       @Test
+      @TestMetadata("InjectingSubclassWithNoMembersButParentsDo.kt")
+      public void testInjectingSubclassWithNoMembersButParentsDo() {
+        runTest("compiler-tests/src/test/data/box/inject/member/InjectingSubclassWithNoMembersButParentsDo.kt");
+      }
+
+      @Test
       @TestMetadata("MemberInjectorRequest.kt")
       public void testMemberInjectorRequest() {
         runTest("compiler-tests/src/test/data/box/inject/member/MemberInjectorRequest.kt");
@@ -508,6 +514,18 @@ public class BoxTestGenerated extends AbstractBoxTest {
       @TestMetadata("MemberInjectorRequestMultipleLocations.kt")
       public void testMemberInjectorRequestMultipleLocations() {
         runTest("compiler-tests/src/test/data/box/inject/member/MemberInjectorRequestMultipleLocations.kt");
+      }
+
+      @Test
+      @TestMetadata("MemberInjectsInMultibinding.kt")
+      public void testMemberInjectsInMultibinding() {
+        runTest("compiler-tests/src/test/data/box/inject/member/MemberInjectsInMultibinding.kt");
+      }
+
+      @Test
+      @TestMetadata("MultiInheritanceMemberInject.kt")
+      public void testMultiInheritanceMemberInject() {
+        runTest("compiler-tests/src/test/data/box/inject/member/MultiInheritanceMemberInject.kt");
       }
     }
   }