+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Changelog
 - **Fix:** Propagate property annotations for `@Provides`-properties, previously only the accessor function annotations were being included.
 - **Fix:** Propagate class annotations for `@Inject`-annotated constructors to factory class mirror functions, previously only the constructor's annotations were being included.
 - **Fix:** Fix dispatch receiver for `DelegateFactory` fields when `chunkFieldInits` is enabled.
+- **Fix:** Fix compilation error for members-injected classes with no direct, but only inherited `@Inject` attributes.
+- **Fix:** Always look up member injectors of ancestor classes of classes member-injected by graphs (sorry, word soup I know).
 - Add a `ViewModel` assisted injection example to `compose-navigation-app` sample.
 - Small improvements to the doc site (404 page, favicon, etc.)